+++ packages/diagram_generator/lib/main.dart
@@ -41,8 +41,10 @@ Future<void> main(List<String> args) async {
   DiagramFlutterBinding.ensureInitialized();
   late final List<String> arguments;
   if (platform.isAndroid) {
-    arguments = window.defaultRouteName.length > 5
-        ? Uri.decodeComponent(window.defaultRouteName.substring(5)).split(' ')
+    arguments = PlatformDispatcher.instance.defaultRouteName.length > 5
+        ? Uri.decodeComponent(
+                PlatformDispatcher.instance.defaultRouteName.substring(5))
+            .split(' ')
         : <String>[];
   } else {
     arguments = args;
@@ -86,7 +88,7 @@ Future<void> main(List<String> args) async {
 
   final DiagramController controller = DiagramController(
     outputDirectory: outputDirectory,
-    screenDimensions: const Size(1000.0, 1000.0),
+    screenDimensions: const Size(1300.0, 1300.0),
     pixelRatio: 1.0,
   );
 
@@ -157,6 +159,7 @@ Future<void> main(List<String> args) async {
               category: step.category,
               name: diagram.name,
               start: diagram.startAt,
+              videoFormat: diagram.videoFormat,
             );
           } else {
             await controller.drawDiagramToFile(