+++ packages/diagram_viewer/android/app/src/main/kotlin/com/example/diagram_viewer/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.example.diagram_viewer
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}