+++ packages/diagrams/lib/src/curve.dart
@@ -165,7 +165,10 @@ class CurveDescription extends CustomPainter {
       canvas.drawPath(valueMarker, _valueMarkerPaint);
     }
     final Path graph = Path()..moveTo(area.left, area.bottom);
-    final double stepSize = 1.0 / (area.width * ui.window.devicePixelRatio);
+    final double stepSize = 1.0 /
+        (area.width *
+            (ui.PlatformDispatcher.instance.implicitView?.devicePixelRatio ??
+                1.0));
     for (double t = 0.0;
         t <= (position == 0.0 ? 1.0 : position);
         t += stepSize) {
@@ -599,6 +602,10 @@ class CurveDiagramStep extends DiagramStep {
         name: 'fast_linear_to_slow_ease_in',
         caption: 'Curves.fastLinearToSlowEaseIn',
         curve: Curves.fastLinearToSlowEaseIn),
+    const CurveDiagram(
+        name: 'fast_ease_in_to_slow_ease_out',
+        caption: 'Curves.fastEaseInToSlowEaseOut',
+        curve: Curves.fastEaseInToSlowEaseOut)
   ];
 
   @override